+++ CHANGELOG.md
@@ -15,4 +15,5 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
-- fixes/1200732308074785 - Fix collections by adding getChainItems
+- fix/1200732308074785 - Fix collections by adding getChainItems
+- fix/1201692512287210 - Add $prefixDirectory parameter to EnvFileParser::parse